decompose code
