+++ src/utils/create-new-project/index.js
@@ -2,7 +2,7 @@ const path = require('path')
 const fs = require('fs-extra')
 const execa = require('execa')
 const addConfig = require('../add-config')
-const setupGit = require('../setup-git')
+const setupGit = require('./setup-git')
 const log = require('../log')
 const scripts = require('../scripts')
 const createNewApp = require('./create-new-app')
