+++ test/utils/add-plugin.test.js
@@ -1,12 +1,14 @@
 /* eslint-env mocha */
 
-const chai = require('chai')
-const path = require('path')
-const fs = require('fs-extra')
 const os = require('os')
+const path = require('path')
+const chai = require('chai')
+const fs = require('fs-extra')
 const proxyquire = require('proxyquire')
 
 const modulePath = '../../src/utils/add-plugin'
+const addNpmPluginModulePath = path.join(modulePath, 'add-npm-plugin')
+
 const expect = chai.expect
 const temp = os.tmpdir()
 
@@ -28,8 +30,11 @@ describe('utils/add-plugin', function () {
   })
 
   it('should add a plugin to an app project', async () => {
-    const addPlugin = proxyquire(modulePath, {
+    const addNpmPlugin = proxyquire(addNpmPluginModulePath, {
       'latest-version': async () => '1.0.0'
+    })
+    const addPlugin = proxyquire(modulePath, {
+      './add-npm-plugin': addNpmPlugin
     })
     await addPlugin(appPath, 'provider', 'test-provider', defaultOptions)
 
@@ -50,9 +55,13 @@ describe('utils/add-plugin', function () {
   })
 
   it('should add a plugin published as a scoped module', async () => {
-    const addPlugin = proxyquire(modulePath, {
+    const addNpmPlugin = proxyquire(addNpmPluginModulePath, {
       'latest-version': async () => '1.0.0'
     })
+    const addPlugin = proxyquire(modulePath, {
+      './add-npm-plugin': addNpmPlugin
+    })
+
     await addPlugin(appPath, 'provider', '@koop/test-provider', defaultOptions)
 
     const plugins = await fs.readFile(path.join(appPath, 'src', 'plugins.js'), 'utf-8')
@@ -72,8 +81,11 @@ describe('utils/add-plugin', function () {
   })
 
   it('should add a plugin published with a version number', async () => {
-    const addPlugin = proxyquire(modulePath, {
+    const addNpmPlugin = proxyquire(addNpmPluginModulePath, {
       'latest-version': async () => '3.2.1'
+    })
+    const addPlugin = proxyquire(modulePath, {
+      './add-npm-plugin': addNpmPlugin
     })
     await addPlugin(appPath, 'provider', '@koop/test-provider@3.2.0', defaultOptions)
 
@@ -94,8 +106,11 @@ describe('utils/add-plugin', function () {
   })
 
   it('should add plugin config if provided', async () => {
-    const addPlugin = proxyquire(modulePath, {
+    const addNpmPlugin = proxyquire(addNpmPluginModulePath, {
       'latest-version': async () => '3.2.1'
+    })
+    const addPlugin = proxyquire(modulePath, {
+      './add-npm-plugin': addNpmPlugin
     })
     await addPlugin(
       appPath,
@@ -114,8 +129,11 @@ describe('utils/add-plugin', function () {
   })
 
   it('should append the plugin config to the app config if specified', async () => {
-    const addPlugin = proxyquire(modulePath, {
+    const addNpmPlugin = proxyquire(addNpmPluginModulePath, {
       'latest-version': async () => '3.2.1'
+    })
+    const addPlugin = proxyquire(modulePath, {
+      './add-npm-plugin': addNpmPlugin
     })
     await addPlugin(
       appPath,
@@ -133,8 +151,11 @@ describe('utils/add-plugin', function () {
   })
 
   it('should add the output plugin to the output list in the project', async () => {
-    const addPlugin = proxyquire(modulePath, {
+    const addNpmPlugin = proxyquire(addNpmPluginModulePath, {
       'latest-version': async () => '3.2.1'
+    })
+    const addPlugin = proxyquire(modulePath, {
+      './add-npm-plugin': addNpmPlugin
     })
     await addPlugin(appPath, 'output', '@koop/output-tile', defaultOptions)
 
@@ -152,8 +173,11 @@ describe('utils/add-plugin', function () {
   })
 
   it('should add the auth plugin to the auth list in the project', async () => {
-    const addPlugin = proxyquire(modulePath, {
+    const addNpmPlugin = proxyquire(addNpmPluginModulePath, {
       'latest-version': async () => '3.2.1'
+    })
+    const addPlugin = proxyquire(modulePath, {
+      './add-npm-plugin': addNpmPlugin
     })
     await addPlugin(appPath, 'auth', '@koop/my-auth', defaultOptions)
 
@@ -171,8 +195,11 @@ describe('utils/add-plugin', function () {
   })
 
   it('should add the plugin options to the plugin list', async () => {
-    const addPlugin = proxyquire(modulePath, {
+    const addNpmPlugin = proxyquire(addNpmPluginModulePath, {
       'latest-version': async () => '3.2.1'
+    })
+    const addPlugin = proxyquire(modulePath, {
+      './add-npm-plugin': addNpmPlugin
     })
     await addPlugin(
       appPath,
